+++ homeassistant/components/evohome/manifest.json
@@ -6,5 +6,5 @@
     "evohomeclient==0.3.2"
   ],
   "dependencies": [],
-  "codeowners": []
+  "codeowners": ["@zxdavb"]
 }
